refactor(mypage): 섭취한 당 기록 보기 응답구조 변경

+++ sweet-choice/src/main/java/com/codeon/sweet_choice/service/FoodService.java
@@ -83,7 +83,7 @@ public class FoodService {
                 .recordId(saved.getRecordId())
                 .foodName(food.getFoodName())
                 .count(saved.getRecordCount())
-                .calcSugar(food.getTotalSugar() * saved.getRecordCount())
+                .calcSugar((double) (food.getTotalSugar() * saved.getRecordCount()))
                 .recordDate(saved.getRecordDate())
                 .build();
     }
